add Payment Component
