+++ client/src/components/Payment.jsx
@@ -1,0 +1,95 @@
+import React, { useState } from "react";
+import { loadStripe } from "@stripe/stripe-js";
+import { Elements, CardElement, useStripe, useElements } from "@stripe/react-stripe-js";
+
+import axios from "axios";
+
+const stripePromise = loadStripe("pk_test_51Lj7u9Dukzk6GqlmMVPMHwx3lemmPxbgy32HRrVczr7jnBiVC4MGYPQBzrrQz99uhWYB1KLstQ6PKDBJVDyROyXz00qQ34D93u");
+
+const CheckoutForm = () => {
+  const stripe = useStripe();
+  const elements = useElements();
+
+  const [loading, setLoading] = useState(false);
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+
+    const { error, paymentMethod } = await stripe.createPaymentMethod({
+      type:"card",
+      card: elements.getElement(CardElement),
+    });
+    setLoading(true);
+
+    if (!error) {
+      console.log(paymentMethod)
+      const { id } = paymentMethod;
+      try {
+        const { data } = await axios.post(
+          "/api/checkout",
+          {
+            id,
+            amount: 10000, //cents
+          }
+        );
+        console.log(data);
+
+        elements.getElement(CardElement).clear();
+      } catch (error) {
+        console.log(error);
+      }
+      setLoading(false);
+    }
+  };
+
+  console.log(!stripe || loading);
+
+  return (
+    <form className="card card-body" onSubmit={handleSubmit}>
+      {/* Product Information */}
+      <img
+        src="https://www.corsair.com/medias/sys_master/images/images/h80/hdd/9029904465950/-CH-9109011-ES-Gallery-K70-RGB-MK2-01.png"
+        alt="Corsair Gaming Keyboard RGB"
+        className="w-20 h-20"
+      />
+
+      <h3 className="text-center my-2">Price: 100$</h3>
+
+      {/* User Card Input */}
+      <div className="form-group">
+        <CardElement />
+      </div>
+
+      <button disabled={!stripe} className="btn btn-success">
+        {loading ? (
+          <div className="spinner-border text-light" role="status">
+            <span className="sr-only">Loading...</span>
+          </div>
+        ) : (
+          "Buy"
+        )}
+      </button>
+    </form>
+  );
+};
+
+function Payment() {
+  
+    return (
+     <div>
+  {/* stripe */}
+  <Elements stripe={stripePromise}>
+  <div className='grid gap-4 m-4 justify-items-center'>
+    <div>
+      <div className='w-80 border'>
+        <CheckoutForm />
+      </div>
+    </div>
+  </div>
+  </Elements>
+      </div>
+      
+    );
+  }
+  
+  export default Payment;
